Monster2 Hp Bar commit - error

+++ RPG_Project/Source/RPG_Project/AI/MonsterData.h
@@ -17,7 +17,7 @@ struct RPG_PROJECT_API FMonsterData : public FTableRowBase
 	~FMonsterData() {}
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
-		int HP;
+		float HP;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
 		int ATT;
